requirements.txt
Flask
Flask-CORS
requests
python-dotenv
gunicorn
Pillow
pystray
plyer
webview
python-binance